+++ tilebox-datasets/tilebox/datasets/query/__init__.py
@@ -1,3 +1,4 @@
+from tilebox.datasets.query.expression import Expression, Field, field
 from tilebox.datasets.query.time_interval import (
     TimeInterval,
     TimeIntervalLike,
@@ -5,4 +6,12 @@ from tilebox.datasets.query.time_interval import (
     timestamp_to_datetime,
 )
 
-__all__ = ["TimeInterval", "TimeIntervalLike", "datetime_to_timestamp", "timestamp_to_datetime"]
+__all__ = [
+    "Expression",
+    "Field",
+    "TimeInterval",
+    "TimeIntervalLike",
+    "datetime_to_timestamp",
+    "field",
+    "timestamp_to_datetime",
+]